backend/requirements.txt
absl-py==2.0.0
paddlepaddle==2.6.1
paddleocr==2.7.3
rapidfuzz==3.6.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.56.0
anyio==4.3.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blis==1.0.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.0.0
comm==0.1.4
confection==0.1.5
contourpy==1.1.1
cycler==0.11.0
cymem==2.0.8
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
easyocr==1.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==1.2.0
fastapi==0.115.14
fastjsonschema==2.19.1
filelock==3.18.0
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.59.3
h11==0.14.0
h5py==3.12.1
httpcore==1.0.3
httptools==0.6.4
httpx==0.26.0
idna==3.6
imageio==2.37.0
ipykernel==6.25.2
ipython==8.15.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.3
jiter==0.10.0
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
keras==2.15.0
kiwisolver==1.4.5
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
libclang==16.0.6
marisa-trie==1.2.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
mss==10.0.0
multidict==6.6.3
murmurhash==1.0.10
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
ninja==1.11.1.4
notebook==7.1.0
notebook_shim==0.2.4
numpy==2.0.2
oauthlib==3.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.3.0
platformdirs==3.10.0
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.39
propcache==0.3.2
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post6
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.16.1
PyMuPDF==1.24.13
pyparsing==3.1.1
python-bidi==0.6.6
python-dateutil==2.8.2
python-dotenv==1.1.1
python-engineio==4.12.2
python-json-logger==2.0.7
python-socketio==5.13.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.18.0
rsa==4.9
scikit-image==0.25.2
scikit-learn==1.3.2
scipy==1.16.0
seaborn==0.13.0
Send2Trash==1.8.2
shapely==2.1.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smart-open==7.0.5
sniffio==1.3.0
soupsieve==2.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.2
starlette==0.46.2
sympy==1.14.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.3.0
terminado==0.18.0
thinc==8.3.2
threadpoolctl==3.2.0
tifffile==2025.6.11
tinycss2==1.2.1
torch==2.7.1
torchvision==0.22.1
tornado==6.3.3
tqdm==4.67.0
traitlets==5.10.1
typer==0.13.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.14.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.35.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.6
weasel==0.4.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==15.0.1
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
yarl==1.20.1
